Minor Bug Fix
Fixed
A bug from the start of the project causing hex strings to only support 32 bit numbers (it was a face palm moment i casted the value into 32 bit smh)

+++ kernel/cstdlib/stdio/LouPrint.c
@@ -95,9 +95,9 @@ int LouPrint(char* format, ...) {
                     break;
                 }
                 case 'h': {
-                    int64_t num = va_arg(args, int32_t);
+                    int64_t num = va_arg(args, uint64_t);
                     char hexString[21];
-                    uintToHexString((uint32_t)num, hexString);
+                    uintToHexString((uint64_t)num, hexString);
                     char* p = hexString;
                     while (*p  != '\0') {
                         VgaPutCharecterRgb(*p++, DebugWindow, 0, 255, 0);
